+++ packages/client/src/actions/index.js
@@ -89,11 +89,6 @@ export default config => {
     return add(parentPath, node, { index })
   }
 
-  let shallowCloneNode = node => ({
-    ...node,
-    children: [...node.children],
-  })
-
   let { wrapInGroup } = wrap(config, {mutate, replace, add})
 
   let move = (path, { path: targetPath, index: targetIndex } = {}) => {
@@ -122,7 +117,6 @@ export default config => {
 
   let pauseNested = path => mutateNested(path, { paused: true })
   let unpauseNested = path => mutateNested(path, { paused: false })
-  let setPausedNested = (path, paused) => mutateNested(path, { paused })
 
   let nodeLeaves = _.flow(getNode, Tree.leaves)
   let isPausedNested = _.flow(
@@ -143,6 +137,5 @@ export default config => {
     isPausedNested,
     pauseNested,
     unpauseNested,
-    setPausedNested,
   }
 }